requirements.txt
aiohttp==3.7.4.post0
astroid==2.8.0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
colorama==0.4.4
discord==1.7.3
discord.py==1.7.3
idna==3.2
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
multidict==5.1.0
numpy==1.21.2
platformdirs==2.4.0
pycparser==2.20
pylint==2.11.1
PyNaCl==1.4.0
python-dotenv==0.19.0
requests==2.26.0
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
wrapt==1.12.1
yarl==1.6.3
youtube-dl==2021.6.6
